* bugfix repetition() method * rewrite nega-max using https://en.wikipedia.org/wiki/Negamax * pack eval in i8 in the hashtable * don't call repetition in positions with sente move

+++ src/main.rs
@@ -111,7 +111,7 @@ fn process_file(
     s.reset();
     match s.iterative_search(&mut pos, 1, depth) {
       Some(res) => {
-        if res < depth as i16 {
+        if res < depth as i32 {
           warn!(
             "Found faster mate in {} move(s). Test #{}, sfen: {}",
             res,
